drop ColorBuilder<T>
